Feature: Familiar/Pet improvements
You can now operate the orb interface (Need to find right var config to update the options), rest features work. Until I can get the var config found I cant finish the combat related stuff; won't be long

+++ src/com/rs/game/player/PlayerDetails.java
@@ -151,7 +151,7 @@ public final class PlayerDetails {
 	private byte tradeStatus;
 	private byte assistStatus;
 
-	private byte summoningLeftClickOption;
+	private int summoningLeftClickOption;
 	
 	// Used for storing recent ips and password
 	private ObjectArrayList<String> passwordList = new ObjectArrayList<String>();
